+++ src/main/java/com/event/university/entity/BinhLuanBaiViet.java
@@ -17,7 +17,7 @@ public class BinhLuanBaiViet {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 
 	@Column(name = "thoigian", nullable = false)
 	private LocalDateTime thoiGian;
@@ -36,11 +36,11 @@ public class BinhLuanBaiViet {
 	public BinhLuanBaiViet() {
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
