+++ src/main/java/com/event/university/repository/KhoaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.event.university.entity.Khoa;
 
-public interface KhoaRepository extends JpaRepository<Khoa, Long> {
+public interface KhoaRepository extends JpaRepository<Khoa, Integer> {
 }
