+++ src/main/java/com/event/university/repository/LopRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.event.university.entity.Lop;
 
-public interface LopRepository extends JpaRepository<Lop, Long> {
-	List<Lop> findByNganhId(Long nganhId);
+public interface LopRepository extends JpaRepository<Lop, Integer> {
+	List<Lop> findByNganhId(Integer nganhId);
 }
