+++ src/main/java/com/event/university/repository/VaiTroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.event.university.entity.VaiTro;
 
-public interface VaiTroRepository extends JpaRepository<VaiTro, Long> {
+public interface VaiTroRepository extends JpaRepository<VaiTro, Integer> {
 	VaiTro findByBiDanh(String biDanh);
 }
